requirements.txt
# Core dependencies
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0

# Utilities
typing-extensions==4.8.0
werkzeug==2.3.7

# Production (optional)
# gunicorn==21.2.0